requirements.txt
fritzconnection==1.9.1
python-dotenv==0.21.0
pyVoIP==1.5.6
